move asset hub definition tests
