+++ runtime/common/src/lib.rs
@@ -162,28 +162,3 @@ pub mod parachains {
 	}
 }
 
-#[cfg(test)]
-mod tests {
-	use crate::parachains::asset_hub;
-	use xcm::{
-		latest::prelude::{PalletInstance, Parachain},
-		prelude::GeneralIndex,
-	};
-
-	#[test]
-	fn test_PINK() {
-		let pink_loc = asset_hub::PINK_location();
-		let mut junctions = pink_loc.interior().into_iter();
-
-		assert_eq!(junctions.next(), Some(&Parachain(asset_hub::PARA_ID)));
-		assert_eq!(junctions.next(), Some(&PalletInstance(asset_hub::ASSET_PALLET_INDEX)));
-		assert_eq!(junctions.next(), Some(&GeneralIndex(asset_hub::PINK_ASSET_ID)));
-		assert_eq!(junctions.next(), None);
-	}
-
-	#[test]
-	fn test_constants() {
-		let expected_USDT_value = 1984;
-		assert_eq!(asset_hub::USDT_ASSET_ID, expected_USDT_value);
-	}
-}
